+++ src/main/kotlin/com/github/insanusmokrassar/IObjectK/extensions/IInputObject.kt
@@ -3,12 +3,17 @@ package com.github.insanusmokrassar.IObjectK.extensions
 import com.github.insanusmokrassar.IObjectK.exceptions.ReadException
 import com.github.insanusmokrassar.IObjectK.interfaces.IInputObject
 import com.github.insanusmokrassar.IObjectK.interfaces.has
+import com.github.insanusmokrassar.IObjectK.realisations.StandardIInputObjectIterator
 
 /**
  * Creating map to IInputObject
  */
 fun <K, V> IInputObject<K, V>.asMap(): Map<K, V> {
     return IInputObjectMap(this)
+}
+
+fun <K, V> IInputObject<K, V>.iterator(): Iterator<Pair<K, V>> {
+    return StandardIInputObjectIterator(this)
 }
 
 /**