+++ CHANGELOG.md
@@ -47,3 +47,8 @@ extension
 * `IInputObject` [now is `Iterable<Pair<K, V>>`](src/main/kotlin/com/github/insanusmokrassar/IObjectK/interfaces/IInputObject.kt#L6).
 You can use [StandardIInputObjectIterator](src/main/kotlin/com/github/insanusmokrassar/IObjectK/realisations/StandardIInputObjectIterator.kt)
 if you need not some special extends.
+
+## v1.9
+
+* Remove implementation of Iterable for IInputObject's (you can use [StandardIInputObjectIterator](src/main/kotlin/com/github/insanusmokrassar/IObjectK/realisations/StandardIInputObjectIterator.kt)
+or [IInputObject#iterator](src/main/kotlin/com/github/insanusmokrassar/IObjectK/extensions/IInputObject.kt#L15))